# docs/openapi3.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Konveyor Hub API",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.3.z"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/applications/{id}/stakeholders": {
      "patch": {
        "tags": [
          "applications"
        ],
        "summary": "Update the owner and contributors of an Application.",
        "description": "Update the owner and contributors of an Application.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Application ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Application stakeholders",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api.Stakeholders"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "application"
      }
    },
...
    "/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "List all settings.",
        "description": "List all settings.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Setting"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Create a setting.",
        "description": "Create a setting.",
        "requestBody": {
          "description": "Setting data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Setting"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Setting"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "setting"
      }
    },
    "/settings/{key}": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get a setting by its key.",
        "description": "Get a setting by its key.",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Setting"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update a setting.",
        "description": "Update a setting.",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        }
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Create a setting.",
        "description": "Create a setting.",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Setting value",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Setting"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "setting"
      },
      "delete": {
        "tags": [
          "settings"
        ],
        "summary": "Delete a setting.",
        "description": "Delete a setting.",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        }
      }
    },
...
    "/stakeholders": {
      "get": {
        "tags": [
          "stakeholders"
        ],
        "summary": "List all stakeholders.",
        "description": "List all stakeholders.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Stakeholder"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "stakeholders"
        ],
        "summary": "Create a stakeholder.",
        "description": "Create a stakeholder.",
        "requestBody": {
          "description": "Stakeholder data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Stakeholder"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Stakeholder"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "stakeholder"
      }
    },
    "/stakeholders/{id}": {
      "get": {
        "tags": [
          "stakeholders"
        ],
        "summary": "Get a stakeholder by ID.",
        "description": "Get a stakeholder by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Stakeholder ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Stakeholder"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "stakeholders"
        ],
        "summary": "Update a stakeholder.",
        "description": "Update a stakeholder.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Stakeholder ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Stakeholder data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Stakeholder"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "stakeholder"
      },
      "delete": {
        "tags": [
          "stakeholders"
        ],
        "summary": "Delete a stakeholder.",
        "description": "Delete a stakeholder.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Stakeholder ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "api.Stakeholders": {
        "type": "object",
        "properties": {
          "contributors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_internal_api.Ref"
            }
          },
          "owner": {
            "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_internal_api.Ref"
          }
        }
      },
...
      "github_com_konveyor_tackle2-hub_api.Ref": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "github_com_konveyor_tackle2-hub_api.Setting": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "object"
          }
        }
      },
      "github_com_konveyor_tackle2-hub_api.Stakeholder": {
        "required": [
          "email",
          "name"
        ],
        "type": "object",
        "properties": {
          "businessServices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
            }
          },
          "contributes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
            }
          },
          "createTime": {
            "type": "string"
          },
          "createUser": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "jobFunction": {
            "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
          },
          "migrationWaves": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
            }
          },
          "name": {
            "type": "string"
          },
          "owns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
            }
          },
          "stakeholderGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/github_com_konveyor_tackle2-hub_api.Ref"
            }
          },
          "updateUser": {
            "type": "string"
          }
        }
      },
...
      "github_com_konveyor_tackle2-hub_internal_api.Ref": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}